implement authorize user to edit and delete the question using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -67,6 +67,9 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if(\Gate::denies('update-question', $question)) {
+            abort(403, "access denied");
+        };
         return view('questions.edit', compact('question'));
     }
 
@@ -79,6 +82,9 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question', $question)) {
+            abort(403, "access denied");
+        };
         $question->update($request->only('title', 'body'));
 
         return redirect()->route('questions.index')->with('update', 'Your question has been updated');
@@ -92,6 +98,9 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question', $question)) {
+            abort(403, "access denied");
+        };
         $question->delete();
         
         return redirect()->route('questions.index')->with('delete', 'Your question has been deleted');
